SC Refactoring: Split block storage operations into dedicated components
* Introduced separate types (`batchOp`, `batchCountOp`, `stateRootOp`, etc.) to represent isolated sets of operations on different data structures: blocks, batches, secondary indices, etc;
* Reduced the size of `block_storage.go` file from ~1100 lines to ~600;
* No functional changes introduced
